Fix for linting issues in V2, Some of the issue are disabled bacuase those are related with sql alchemy, schema inputs, graphql methods structure etc.(like too-few-public-methods, too-many-arguments).
