+++ app/.pylintrc
@@ -4,7 +4,7 @@ init-hook='from pylint.config import find_pylintrc; import os, sys; sys.path.app
 # A comma-separated list of package or module names from where C extensions may
 # be loaded. Extensions are loading into the active Python interpreter and may
 # run arbitrary code.
-extension-pkg-allow-list='pydantic'
+extension-pkg-allow-list=pydantic
 
 # A comma-separated list of package or module names from where C extensions may
 # be loaded. Extensions are loading into the active Python interpreter and may
