+++ app/crud/nlp_tokenization.py
@@ -9,7 +9,6 @@ import db_models
 from custom_exceptions import NotAvailableException
 
 
-
 def build_memory_trie(translation_memory):
     '''form a trie from a list of known tokens in a source language, to be used for tokenization'''
     memory_trie = pygtrie.StringTrie()
