+++ app/graphql_api/types.py
@@ -2,7 +2,7 @@
 import graphene
 from graphene.types import Scalar
 
-#pylint: disable=too-few-public-methods
+#Data clases have few public methods methods
 class Metadata(Scalar):
     '''metadata representing JSON'''
     @staticmethod
@@ -10,12 +10,12 @@ class Metadata(Scalar):
         '''process the response'''
         return dt_
 
-class ContentType(graphene.ObjectType):
+class ContentType(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''output object for content types'''
     contentId = graphene.ID()
     contentType = graphene.String(description="Use values bible, commentary, dictionary etc")
 
-class Language(graphene.ObjectType):
+class Language(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''output object for Language'''
     languageId = graphene.Int()
     language = graphene.String()
@@ -32,7 +32,7 @@ class LicensePermission(graphene.Enum):
     private = "Private_use"
 
 
-class License(graphene.ObjectType):
+class License(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Return object of licenses'''
     name = graphene.String()
     code = graphene.String()
@@ -40,7 +40,7 @@ class License(graphene.ObjectType):
     permissions = graphene.List(LicensePermission)
     active = graphene.Boolean()
 
-class Version(graphene.ObjectType):
+class Version(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Return object of version'''
     versionId = graphene.ID()
     versionAbbreviation = graphene.String()
@@ -48,7 +48,7 @@ class Version(graphene.ObjectType):
     revision = graphene.Int()
     metaData = Metadata()
 
-class Source(graphene.ObjectType):
+class Source(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Return object of source'''
     sourceName = graphene.String()
     contentType = graphene.Field(ContentType)
@@ -60,13 +60,13 @@ class Source(graphene.ObjectType):
     active = graphene.Boolean()
 
 
-class BibleBook(graphene.ObjectType):
+class BibleBook(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''response object of Bible book'''
     bookId = graphene.ID()
     bookName = graphene.String()
     bookCode = graphene.String()
 
-class AudioBible(graphene.ObjectType):
+class AudioBible(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''output object for AudioBible'''
     name = graphene.String()
     url = graphene.String()
@@ -74,7 +74,7 @@ class AudioBible(graphene.ObjectType):
     format = graphene.String()
     active = graphene.Boolean()
 
-class BibleContent(graphene.ObjectType):
+class BibleContent(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''output object for Biblecontent'''
     book = graphene.Field(BibleBook)
     USFM = graphene.String()
@@ -82,14 +82,14 @@ class BibleContent(graphene.ObjectType):
     audio = graphene.Field(AudioBible)
     active = graphene.Boolean()
 
-class Versification(graphene.ObjectType):
+class Versification(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Format of versification response'''
     maxVerses = Metadata()
     mappedVerses = Metadata()
     excludedVerses = Metadata()
     partialVerses = Metadata()
 
-class Reference(graphene.ObjectType):
+class Reference(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Response object of bible refernce'''
     bible = graphene.String()
     book = graphene.String()
@@ -149,20 +149,20 @@ class Commentary(graphene.ObjectType):
         '''resolver'''
         return parent.book
 
-class DictionaryWord(graphene.ObjectType):
+class DictionaryWord(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Response object for dictionary word'''
     word = graphene.String()
     details = Metadata()
     active = graphene.Boolean()
 
-class Infographic(graphene.ObjectType):
+class Infographic(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Response for Infographics'''
     book = graphene.Field(BibleBook)
     title = graphene.String()
     infographicLink = graphene.String()
     active = graphene.Boolean()
 
-class BibleVideo(graphene.ObjectType):
+class BibleVideo(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Response for BibleVideo'''
     title = graphene.String()
     books = graphene.List(graphene.String)
@@ -179,7 +179,7 @@ class TranslationDocumentType(graphene.Enum):
     CSV = 'csv'
     JSON = 'alignment-json'
 
-class ProjectUser(graphene.ObjectType):
+class ProjectUser(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Input object for AgMT user update'''
     project_id = graphene.ID()
     userId = graphene.Int()
@@ -187,7 +187,7 @@ class ProjectUser(graphene.ObjectType):
     metaData = Metadata()
     active = graphene.Boolean()
 
-class TranslationProject(graphene.ObjectType):
+class TranslationProject(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Output object for project creation'''
     projectId = graphene.ID()
     projectName = graphene.String()
@@ -198,45 +198,45 @@ class TranslationProject(graphene.ObjectType):
     metaData = Metadata()
     active = graphene.Boolean()
 
-class TokenOccurence(graphene.ObjectType):
+class TokenOccurence(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Object for token occurence'''
     sentenceId = graphene.ID()
     offset = graphene.List(graphene.Int)
 
-class Token(graphene.ObjectType):
+class Token(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Response object for token'''
     token = graphene.String()
     occurrences = graphene.List(TokenOccurence)
     translations = Metadata(name="translationSuggestions")
     metaData = Metadata()
 
-class TokenTranslation(graphene.ObjectType):
+class TokenTranslation(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''For translation/draft of a specific token'''
     token = graphene.String()
     translation = graphene.String()
     occurrence = graphene.Field(TokenOccurence)
     status = graphene.String()
 
-class Sentence(graphene.ObjectType):
+class Sentence(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Response object for sentences and plain-text draft'''
     sentenceId = graphene.ID()
     sentence = graphene.String()
     draft = graphene.String()
     draftMeta = Metadata()
 
-class Progress(graphene.ObjectType):
+class Progress(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Response object for AgMT project progress'''
     confirmed = graphene.Float()
     suggestion = graphene.Float()
     untranslated = graphene.Float()
 
-class Gloss(graphene.ObjectType):
+class Gloss(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Output object for translation memory or gloss'''
     token = graphene.String()
     translations = Metadata()
     metaData = Metadata()
 
-class Suggestion(graphene.ObjectType):
+class Suggestion(graphene.ObjectType):#pylint: disable=too-few-public-methods
     '''Response object for suggestion'''
     suggestion = graphene.String()
     score = graphene.Float()
@@ -310,3 +310,348 @@ class SugggestTranslationInput(graphene.InputObjectType):
     stopwords = graphene.Field(Stopwords)
     punctuations = graphene.List(graphene.String,\
         description="""List [ ",", "\"", "!", ".", ":", ";", "\n""]""")
+
+################### Graphql mutation types ###########################
+class InputAddLang(graphene.InputObjectType):
+    """ADD Language Input"""
+    language = graphene.String(required=True)
+    code = graphene.String(required=True,\
+    description="language code as per bcp47(usually 2 letter code)")
+    scriptDirection = graphene.String()
+    metaData = graphene.JSONString(description="Expecting a dictionary Type")
+
+class InputUpdateLang(graphene.InputObjectType):
+    """ update Language Input """
+    languageId = graphene.Int(required=True)
+    language = graphene.String()
+    code = graphene.String(description="language code as per bcp47(usually 2 letter code)")
+    scriptDirection = graphene.String()
+    metaData = graphene.JSONString(description="Expecting a dictionary Type")
+
+class InputContentType(graphene.InputObjectType):
+    """ update Language Input """
+    contentType = graphene.String(required=True,\
+        description="Input object to ceate a new content type : pattern: ^[a-z]+$ :\
+        example: commentary")
+
+class InputAddLicense(graphene.InputObjectType):
+    """Add license Input"""
+    name = graphene.String(required=True)
+    code = graphene.String(required=True,\
+        description="pattern: '^[a-zA-Z0-9\\.\\_\\-]+$'")
+    license = graphene.String(required=True)
+    permissions = graphene.List(LicensePermission, \
+        default_value =["Private_use"],\
+        description="Expecting a list \
+        [ Commercial_use, Modification, Distribution, Patent_use, Private_use ]")
+
+class InputEditLicense(graphene.InputObjectType):
+    """Edit license Input"""
+    name = graphene.String()
+    code = graphene.String(required=True,description=\
+        "pattern: ^[a-zA-Z0-9\\.\\_\\-]+$")
+    license = graphene.String()
+    permissions = graphene.List(LicensePermission, default_value =\
+        ["Private_use"],\
+        description="Expecting a list\
+        [ Commercial_use, Modification, Distribution, Patent_use, Private_use ]")
+    active = graphene.Boolean()
+
+class InputAddVersion(graphene.InputObjectType):
+    """Input for Edit versions"""
+    versionAbbreviation = graphene.String(required=True,\
+        description="pattern: ^[A-Z]+$")
+    versionName = graphene.String(required=True)
+    revision = graphene.Int(default_value = 1)
+    metaData = graphene.JSONString(default_value = None,\
+    description="Expecting a dictionary Type JSON String")
+
+class InputEditVersion(graphene.InputObjectType):
+    """Input for Edit versions"""
+    versionId = graphene.ID(required=True)
+    versionAbbreviation = graphene.String(\
+        description="pattern: ^[A-Z]+$")
+    versionName = graphene.String()
+    revision = graphene.Int()
+    metaData = graphene.JSONString(description="Expecting a dictionary Type JSON String")
+
+class InputAddSource(graphene.InputObjectType):
+    """Add Source Input"""
+    contentType  = graphene.String(required=True)
+    language = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*$")
+    version = graphene.String(required=True,\
+        description="pattern:^[A-Z]+$")
+    revision = graphene.String(default_value = 1,\
+        description="default: 1")
+    year = graphene.Int(required=True)
+    license = graphene.String(default_value = "CC-BY-SA",\
+        description="pattern: ^[a-zA-Z0-9\\.\\_\\-]+$")
+    metaData = graphene.JSONString(description="Expecting a dictionary Type JSON String")
+
+class InputEditSource(graphene.InputObjectType):
+    """Edit Source Input"""
+    sourceName = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    contentType  = graphene.String()
+    language = graphene.String(description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*$")
+    version = graphene.String(description="pattern:^[A-Z]+$")
+    revision = graphene.String(description="default: 1")
+    year = graphene.Int()
+    license = graphene.String(description="pattern: ^[a-zA-Z0-9\\.\\_\\-]+$")
+    metaData = graphene.JSONString(description="Expecting a dictionary Type JSON String")
+    active = graphene.Boolean()
+
+class InputBibleDict(graphene.InputObjectType):
+    """Add Bible Dict"""
+    USFM = graphene.String()
+    JSON = graphene.JSONString(description="Provide JSON structure\
+         obtained from USFM-Grammar or one like that")
+
+class BibleEditDict(graphene.InputObjectType):
+    """bible books inputs"""
+    book_code = graphene.String(
+        description="pattern:^[a-zA-Z1-9][a-zA-Z][a-zA-Z]$")
+    USFM = graphene.String(description="USFM Data")
+    JSON = graphene.JSONString(description="Provide JSON structure obtained\
+        from USFM-Grammar or one like that")
+    active = graphene.Boolean(default_value = True)
+
+class AudioAdddict(graphene.InputObjectType):
+    """audio input"""
+    name = graphene.String(required=True)
+    url = graphene.String(required=True)
+    books = graphene.List(graphene.String,required=True)
+    format = graphene.String(required=True)
+    active = graphene.Boolean(default_value = True)
+
+class AudioEditdict(graphene.InputObjectType):
+    """audio input"""
+    name = graphene.String()
+    url = graphene.String()
+    books = graphene.List(graphene.String,required=True)
+    format = graphene.String()
+    active = graphene.Boolean(default_value = True)
+
+class CommentaryDict(graphene.InputObjectType):
+    """commentary input"""
+    bookCode = graphene.String(required=True)
+    chapter = graphene.Int(required=True)
+    verseStart = graphene.Int()
+    verseEnd = graphene.Int()
+    commentary = graphene.String(required=True)
+    active = graphene.Boolean(default_value = True)
+
+class CommentaryEditDict(graphene.InputObjectType):
+    """commentary Edit input"""
+    bookCode = graphene.String(required=True)
+    chapter = graphene.Int(required=True)
+    verseStart = graphene.Int()
+    verseEnd = graphene.Int()
+    commentary = graphene.String()
+    active = graphene.Boolean(default_value = True)
+
+class InputCreateAGMTProject(graphene.InputObjectType):
+    """CreateAGMTProject Input"""
+    projectName = graphene.String(required=True,\
+        description="example: Hindi Malayalam Gospels")
+    sourceLanguageCode = graphene.String(required=True,\
+        description="pattern:^[a-zA-Z]+(-[a-zA-Z0-9]+)*$")
+    targetLanguageCode = graphene.String(required=True,\
+        description="pattern:^[a-zA-Z]+(-[a-zA-Z0-9]+)*$")
+    documentFormat = graphene.Field(TranslationDocumentType)
+    useDataForLearning = graphene.Boolean()
+    stopwords = graphene.Field(Stopwords)
+    punctuations = graphene.List(graphene.String,\
+        description="""List [ ",", "\"", "!", ".", ":", ";", "\n""]""")
+    active = graphene.Boolean(default_value=True)
+
+class InputSeclectedBooks(graphene.InputObjectType):
+    """List of selected books from an existing bible in the server"""
+    bible = graphene.String(required=True,\
+        description = "pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$\
+        example: hi_IRV_1_bible")
+    books = graphene.List(graphene.String,required=True,\
+        description = "pattern: ^[a-zA-Z1-9][a-zA-Z][a-zA-Z]$\
+        example: [ 'luk', 'jhn' ]")
+
+class InputEditAGMTProject(graphene.InputObjectType):
+    """CreateAGMTProject Input"""
+    projectId = graphene.Int(required=True)
+    projectName = graphene.String(\
+        description="example: Hindi Malayalam Gospels")
+    selectedBooks = graphene.Field(InputSeclectedBooks)
+    uploadedBooks = graphene.List(graphene.String)
+    useDataForLearning = graphene.Boolean()
+    stopwords = graphene.Field(Stopwords)
+    punctuations = graphene.List(graphene.String,\
+        description="""List [ ",", "\"", "!", ".", ":", ";", "\n""]""")
+    active = graphene.Boolean(default_value=True)
+
+class AGMTUserCreateInput(graphene.InputObjectType):
+    """input of AGMT user create"""
+    project_id = graphene.Int(required=True)
+    user_id = graphene.Int(required=True)
+
+class AGMTUserEditInput(graphene.InputObjectType):
+    """input of AGMT user Edit"""
+    project_id = graphene.Int(required=True)
+    userId = graphene.Int(required=True)
+    userRole = graphene.String()
+    metaData = graphene.JSONString()
+    active = graphene.Boolean()
+
+class BibleVideoDict(graphene.InputObjectType):
+    """BibleVideo input"""
+    title = graphene.String(required=True)
+    books = graphene.List(graphene.String,required=True,\
+        description="provide book codes")
+    videoLink = graphene.String(required=True)
+    description = graphene.String(required=True)
+    theme = graphene.String(required=True)
+    active = graphene.Boolean(default_value = True)
+
+class InputAddBibleVideo(graphene.InputObjectType):
+    """Add BibleVideo Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    video_data = graphene.List(BibleVideoDict)
+
+class BibleVideoEditDict(graphene.InputObjectType):
+    """BibleVideo Edit input"""
+    title = graphene.String(required=True)
+    books = graphene.List(graphene.String)
+    videoLink = graphene.String()
+    description = graphene.String()
+    theme = graphene.String()
+    active = graphene.Boolean(default_value = True)
+
+class InputEditBibleVideo(graphene.InputObjectType):
+    """Edit BibleVideo Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    video_data = graphene.List(BibleVideoEditDict)
+
+class DictionaryDict(graphene.InputObjectType):
+    """Dictionary input"""
+    word = graphene.String(required=True)
+    details = graphene.JSONString(description="Expecting a dictionary Type")
+    active = graphene.Boolean(default_value = True)
+
+class InputAddDictionary(graphene.InputObjectType):
+    """Add Dictionary Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    word_list = graphene.List(DictionaryDict)
+
+class DictionaryEditDict(graphene.InputObjectType):
+    """Dictionary input"""
+    word = graphene.String(required=True)
+    details = graphene.JSONString(description="Expecting a dictionary Type")
+    active = graphene.Boolean(default_value = True)
+
+class InputEditDictionary(graphene.InputObjectType):
+    """Add Dictionary Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    word_list = graphene.List(DictionaryEditDict)
+
+class InfographicDict(graphene.InputObjectType):
+    """Infographic input"""
+    bookCode = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z1-9][a-zA-Z][a-zA-Z]$")
+    title = graphene.String(required=True)
+    infographicLink = graphene.String(required=True,\
+        description="Provide valid URL")
+    active = graphene.Boolean(default_value = True)
+
+class InputAddInfographic(graphene.InputObjectType):
+    """Add Infographic Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    data = graphene.List(InfographicDict)
+
+class InfographicEditDict(graphene.InputObjectType):
+    """Infographic Edit input"""
+    bookCode = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z1-9][a-zA-Z][a-zA-Z]$")
+    title = graphene.String(required=True)
+    infographicLink = graphene.String(\
+        description="Provide valid URL")
+    active = graphene.Boolean(default_value = True)
+
+class InputEditInfographic(graphene.InputObjectType):
+    """Edit Infographic Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    data = graphene.List(InfographicEditDict)
+
+class InputApplyToken(graphene.InputObjectType):
+    """Inputs for Aplly Token"""
+    project_id = graphene.Int(required=True)
+    return_drafts = graphene.Boolean(default_value = True)
+    token = graphene.List(TokenUpdate)
+
+class InputAutoTranslation(graphene.InputObjectType):
+    """Auto Translation Suggestion input"""
+    project_id  = graphene.Int(required=True)
+    books = graphene.List(graphene.String)
+    sentence_id_list = graphene.List(graphene.Int,\
+        description="List of sentance id BCV")
+    sentence_id_range = graphene.List(graphene.Int,\
+        description="List of sentance range BCV , 2 values in list")
+    confirm_all = graphene.Boolean(default_value = False)
+
+class InputAddGloss(graphene.InputObjectType):
+    """Add Gloss input"""
+    source_language = graphene.String(required=True,\
+        description="patten:^[a-zA-Z]+(-[a-zA-Z0-9]+)*$")
+    target_language  = graphene.String(required=True,\
+        description="patten:^[a-zA-Z]+(-[a-zA-Z0-9]+)*$")
+    data = graphene.List(GlossInput,required=True)
+
+class InputAddAlignment(graphene.InputObjectType):
+    """Add Alignement input"""
+    source_language = graphene.String(required=True,\
+        description="patten:^[a-zA-Z]+(-[a-zA-Z0-9]+)*$")
+    target_language  = graphene.String(required=True,\
+        description="patten:^[a-zA-Z]+(-[a-zA-Z0-9]+)*$")
+    data = graphene.List(Alignment,required=True)
+
+class InputAddBible(graphene.InputObjectType):
+    """Add Bible Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    books = graphene.List(InputBibleDict,required=True,\
+        description="Must Provide One of the Two USFM or JSON")
+
+class InputEditBible(graphene.InputObjectType):
+    """Edit Bible Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    books = graphene.List(BibleEditDict,\
+        description="Either JSON or USFM should provide")
+
+class InputAddAudioBible(graphene.InputObjectType):
+    """Add Audio Bible Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    audio_data = graphene.List(AudioAdddict)
+
+class InputEditAudioBible(graphene.InputObjectType):
+    """Edit Audio Bible Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    audio_data = graphene.List(AudioEditdict)
+
+class InputAddCommentary(graphene.InputObjectType):
+    """Add commentary Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    commentary_data = graphene.List(CommentaryDict)
+
+class InputEditCommentary(graphene.InputObjectType):
+    """Edit commentary Input"""
+    source_name = graphene.String(required=True,\
+        description="pattern: ^[a-zA-Z]+(-[a-zA-Z0-9]+)*_[A-Z]+_\\w+_[a-z]+$")
+    commentary_data = graphene.List(CommentaryEditDict)
